sales app renamed to dashboard
