+++ dashboard/routing.py
@@ -1,5 +1,5 @@
 from django.urls import re_path
-from sales.consumers import WebsocketBase
+from dashboard.consumers import WebsocketBase
 
 websocket_urlpatterns =[
     re_path("ws/websocket/",WebsocketBase.as_asgi()),
