+++ ecom_manager/urls.py
@@ -19,7 +19,7 @@ from django.urls import path,include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('',include('sales.urls')),
+    path('',include('dashboard.urls')),
     path('authorization/',include('user.urls')),
     path('amazon/',include('amazon.urls'))
 ]
