+++ templates/add_store_form.html
@@ -2,7 +2,7 @@
 {% load static %}
 
 {% block content %}
-<h1>Add amazon store</h1>
+<h1>Add store</h1>
 
 <form class="form" method="POST" enctype="multipart/form-data">
     {% csrf_token %}
