+++ templates/dashboard.html
@@ -7,8 +7,9 @@
 <div class="graph-view">
     <h1>Sales Graph</h1>
 </div>
+
+<!-- Display the dashbaord if the store is added , else we should see the option to add it-->
 <div class="parent-container">
-    <!-- Child items for amazon -->
     <div class="child-container">
         <h2>Amazon Dashboard</h2>
         <p>Current Time: <span id="timeDisplay"></span></p>
