+++ templates/home.html
@@ -37,7 +37,6 @@
             {% include "authorization/auth_form.html" %}
         </div>
     </div>
-
     {% include "footer.html" %}
 </body>
 </html>
